Changed from byte to int

+++ src/Servers/IIS/IIS/src/Core/IISHttpServer.cs
@@ -165,7 +165,7 @@ namespace Microsoft.AspNetCore.Server.IIS.Core
         }
 
         [UnmanagedCallersOnly]
-        private static byte HandleShutdown(IntPtr pvRequestContext)
+        private static int HandleShutdown(IntPtr pvRequestContext)
         {
             IISHttpServer server = null;
             try
